Use more precise terminology re library versions
On reading the old output, I saw `Language: go 1.11.0` which I read as meaning
"version 1.11.0 of Golang".  This is not what the version field is supposed to
mean: it's a version of the library.

This is attempt 1 at rephrasing the Extended Text output format to categorise a
little differently.  I suspect that this particular wording won't be accepted
and that's fine — this highlights the two bits of code which should change and
folks can debate the better phrasing.

With this particular incarnation of a fix, I now see:

    Library: version=1.11.0 language=go