Refactor into a manager struct
When first written it was not clear how nats.go wold be
changed to expose JS features, this is clearer now and
we can do the right thing here.

This refactors away from the previous singleton connection
handling and package level helper functions into a manager
instance that encapsulates its own connection, timeouts
and more.

One application can then use multiple manager instances
and in general the encapsulation is improved for applications
that manage multiple clusters.

Signed-off-by: R.I.Pienaar <rip@devco.net>