Ensure LoadOrNEw does not try to remake existing
These 2 methods were written before we had error codes so we had
no choice but to just handle any error and try to create it.

Now we can tighten this up a bit and only specifically handle the
one error for resource not found, the rest are returned as the
original error.  This should avoid some weird conditions in code
that tries to recover from missing streams