[CHANGED] Chunk size for backup and restore
For snapshot (backup), we should even ideally set the chunk size
to 0 to let the server decide, but since we need it for the
progress bar, set it to 128KB to match what will be the new
default (if not set) in the server.

For restore, set it to 64KB because on NGS for instance, basic
accounts would then not be able to do any restore because the
chunk size would be too big.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>