Allow governors to not leave the governor on completion
This allows a governor to elect not delete his claim to
a slot when completed doing the work.

Enabling this on a governor with a max age of 1 hour will
have the outcome that a specific piece of work can only be
done n times per hour.

If you want to control a cronjob so it runs once only on
multiple machines you would create a governor with this option
set and the governor will stay at capacity long enough for
all other attempts to timeout

Signed-off-by: R.I.Pienaar <rip@devco.net>