Add consumer and stream balancer
Here we add the `balancer` package that allows us to balance stream and
consumer leaders over a set of servers.

The balancer will determine what an even distribution of stream/consumer
leaders are over the set of available servers are, and then trigger leader
elections to rebalance the distribution.

A perfect distribution is not guaranteed after a run.