Store exact server responses
Till now the responses stored per server would be stripped
of their wrapper that holds the ServerInfo structure, but
this is bad practise when building audit tools and just
leaves a chance for losing data.

Further in future we will want to be able to have tools like
"nats server list" operate on these archives and in that case
we would need these envelopes.

So this preserves the enveloped and as a side effect requires
a lot of checks to be updated and also breaks backward compat
with older archives

Signed-off-by: R.I.Pienaar <rip@devco.net>