Add a noexprlang build tag
This module uses `github.com/expr-lang/expr` which, because of the
extensive use of the `reflect` package, disables go's compiler dead code
elimination which can lead to bigger binaries.

This commit adds a `noexprlang` build tag that allows jsm.go users that
do not use expression matching to entirely disable the use of the expr
module so that they can benefit from go's dead code elimination if they
are eligible to outside jsm.go.

References:
- https://golab.io/talks/getting-the-most-out-of-dead-code-elimination
- https://github.com/aarzilli/whydeadcode
- https://github.com/spf13/cobra/pull/1956

Signed-off-by: Sylvain Rabot <sylvain@abstraction.fr>