(#724) Fix balancer when server count is greater than replica count
When a stream or consumer was created with replica count x, but the server count
was greater than x, the balancer could try to rebalance to an invalid server.

The balancer will now check if a server is in the peer group before attempting
to rebalance to it.

This commit also fixes a bug where the balancer could try and rebalance to a
server that was already the leader.

Signed-off-by: Pieter Loubser <ploubser@gmail.com>